requirements-dev.txt
requests==2.28.1
python-dotenv==0.20.0
discord.py==1.7.3